requirements.txt
Django==4.2.7
openai==0.28.1
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==11.2.1
opencv-python==4.11.0.86
pytesseract==0.3.13
numpy==2.3.0 